Moved settings to own file
